codecrafters submit [skip ci]

+++ src/protocol/handler.cpp
@@ -5,7 +5,7 @@
 std::vector<char> build_response(int32_t correlation_id, const std::vector<char> &response_data)
 {
     Writer writer;
-    writer.write_int32(sizeof(correlation_id) + response_data.size());
+    writer.write_int32(static_cast<int32_t>(sizeof(correlation_id) + static_cast<int32_t>(response_data.size())));
     writer.write_int32(correlation_id);
     writer.write_bytes(response_data);
     return writer.data();
